Useless code in _hash func
Some bugs in the _hash func.
- Implicit declaration of variable char.
- hash |= 0 does nothing in terms of converting it to a 32 bit int.  the conversion happened when you shifted by 5 (all bitwise operations convert to 32 bit ints).
 - ECMA Script 2015 encodes array indexes as unsigned 32 bit ints, and so the seemingly useless >>> 0 i've added at the end actually converts to unsigned 32 bit int to ensure no negative values and then mods it by the size of the hashtable once instead of in every iteration. 

I'm not the JS police i just wanted to help out since i'm not sure how many people are using this as a reference. If anyone reading wants to use a hashtable use an external chaining hash table instead of dealing with increasing the size of an array. 
 
cheers :)